requirements.txt
pydantic-ai
fastapi
pydantic-yaml
python-dotenv
loguru
pymongo
uuid_v7